fixing types size
