+++ src/emitter/combine-emitter.ts
@@ -1,7 +1,7 @@
 import { AnyIterable, Mapper } from 'augmentative-iterable';
 import { combineAsync } from '../async';
 import { EventEmitter } from 'events';
-import { FluentEmitOptions } from '../types';
+import { FluentEmitOptions } from '../types-base';
 import { forEmitOf } from '../for-emit-of';
 
 export function combineEmitter<T>(
