+++ src/fluent-async.ts
@@ -1,4 +1,5 @@
-import { FluentAsyncIterable, FluentEmitOptions } from './types';
+import { FluentAsyncIterable } from './types';
+import { FluentEmitOptions } from './types-base';
 import {
   asyncIterableFuncs,
   asyncResolvingFuncs,
