+++ src/recipes/reduce-recipe.ts
@@ -1,6 +1,6 @@
 import { identity } from '../utils';
 import { AnyIterable } from 'augmentative-iterable';
-import { AsyncReducer } from '../types';
+import { AsyncReducer } from '../types-base';
 
 export function reduceRecipe(reduceAndMap: Function) {
   return function <T, R>(
