+++ src/types-internal/compare-provider.ts
@@ -1,4 +1,4 @@
-import { Predicate } from '../types';
+import { Predicate } from '../types-base';
 
 export interface CompareProvider {
   <T>(a: T): Predicate<T>;
